PRSD-NONE: Moves JourneyDataService helper methods into JourneyDataHelper
