+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/AlreadyRegisteredPage.kt
@@ -3,8 +3,8 @@ package uk.gov.communities.prsdb.webapp.forms.pages
 import org.springframework.ui.Model
 import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
+import uk.gov.communities.prsdb.webapp.helpers.JourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.formModels.FormModel
-import uk.gov.communities.prsdb.webapp.services.JourneyDataService
 import kotlin.reflect.KClass
 
 class AlreadyRegisteredPage(
@@ -22,7 +22,7 @@ class AlreadyRegisteredPage(
     ): String {
         model.addAttribute(
             "singleLineAddress",
-            JourneyDataService.getFieldStringValue(journeyData!!, selectedAddressPathSegment, "address"),
+            JourneyDataHelper.getFieldStringValue(journeyData!!, selectedAddressPathSegment, "address"),
         )
 
         return super.populateModelAndGetTemplateName(validator, model, pageData, prevStepUrl)
