+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/helpers/JourneyDataHelper.kt
@@ -1,8 +1,10 @@
 package uk.gov.communities.prsdb.webapp.helpers
 
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
+import uk.gov.communities.prsdb.webapp.forms.journeys.PageData
+import uk.gov.communities.prsdb.webapp.forms.journeys.objectToStringKeyedMap
 import uk.gov.communities.prsdb.webapp.models.dataModels.AddressDataModel
-import uk.gov.communities.prsdb.webapp.services.JourneyDataService
+import java.time.LocalDate
 
 open class JourneyDataHelper {
     companion object {
@@ -11,11 +13,11 @@ open class JourneyDataHelper {
             lookupAddressPathSegment: String,
         ): Pair<String, String>? {
             val houseNameOrNumber =
-                JourneyDataService.getFieldStringValue(journeyData, lookupAddressPathSegment, "houseNameOrNumber")
+                getFieldStringValue(journeyData, lookupAddressPathSegment, "houseNameOrNumber")
                     ?: return null
 
             val postcode =
-                JourneyDataService.getFieldStringValue(journeyData, lookupAddressPathSegment, "postcode")
+                getFieldStringValue(journeyData, lookupAddressPathSegment, "postcode")
                     ?: return null
 
             return Pair(houseNameOrNumber, postcode)
@@ -26,35 +28,35 @@ open class JourneyDataHelper {
             manualAddressPathSegment: String,
         ): AddressDataModel? {
             val addressLineOne =
-                JourneyDataService.getFieldStringValue(
+                getFieldStringValue(
                     journeyData,
                     manualAddressPathSegment,
                     "addressLineOne",
                 ) ?: return null
 
             val townOrCity =
-                JourneyDataService.getFieldStringValue(
+                getFieldStringValue(
                     journeyData,
                     manualAddressPathSegment,
                     "townOrCity",
                 ) ?: return null
 
             val postcode =
-                JourneyDataService.getFieldStringValue(
+                getFieldStringValue(
                     journeyData,
                     manualAddressPathSegment,
                     "postcode",
                 ) ?: return null
 
             val addressLineTwo =
-                JourneyDataService.getFieldStringValue(
+                getFieldStringValue(
                     journeyData,
                     manualAddressPathSegment,
                     "addressLineTwo",
                 )
 
             val county =
-                JourneyDataService.getFieldStringValue(
+                getFieldStringValue(
                     journeyData,
                     manualAddressPathSegment,
                     "county",
@@ -67,6 +69,72 @@ open class JourneyDataHelper {
                 addressLineTwo,
                 county,
             )
+        }
+
+        fun getPageData(
+            journeyData: JourneyData,
+            pageName: String,
+            subPageNumber: Int? = null,
+        ): PageData? {
+            var pageData = objectToStringKeyedMap(journeyData[pageName])
+            if (subPageNumber != null && pageData != null) {
+                pageData = objectToStringKeyedMap(pageData[subPageNumber.toString()])
+            }
+            return pageData
+        }
+
+        fun getFieldStringValue(
+            journeyData: JourneyData,
+            urlPathSegment: String,
+            fieldName: String,
+            subPageNumber: Int? = null,
+        ): String? {
+            val pageData = getPageData(journeyData, urlPathSegment, subPageNumber)
+            return pageData?.get(fieldName)?.toString()
+        }
+
+        fun getFieldIntegerValue(
+            journeyData: JourneyData,
+            urlPathSegment: String,
+            fieldName: String,
+            subPageNumber: Int? = null,
+        ): Int? {
+            val fieldAsString =
+                getFieldStringValue(journeyData, urlPathSegment, fieldName, subPageNumber) ?: return null
+            return fieldAsString.toInt()
+        }
+
+        fun getFieldLocalDateValue(
+            journeyData: JourneyData,
+            urlPathSegment: String,
+            fieldName: String,
+            subPageNumber: Int? = null,
+        ): LocalDate? {
+            val fieldAsString =
+                getFieldStringValue(journeyData, urlPathSegment, fieldName, subPageNumber) ?: return null
+            return fieldAsString.let { LocalDate.parse(fieldAsString) }
+        }
+
+        fun getFieldBooleanValue(
+            journeyData: JourneyData,
+            urlPathSegment: String,
+            fieldName: String,
+            subPageNumber: Int? = null,
+        ): Boolean? {
+            val fieldAsString =
+                getFieldStringValue(journeyData, urlPathSegment, fieldName, subPageNumber) ?: return null
+            return fieldAsString == "true"
+        }
+
+        inline fun <reified E : Enum<E>> getFieldEnumValue(
+            journeyData: JourneyData,
+            urlPathSegment: String,
+            fieldName: String,
+            subPageNumber: Int? = null,
+        ): E? {
+            val fieldAsString =
+                getFieldStringValue(journeyData, urlPathSegment, fieldName, subPageNumber) ?: return null
+            return enumValueOf<E>(fieldAsString)
         }
     }
 }
